Implemented OneDrive Integration. This version manages to login the user successfully. It also displays the files in the OneDrive. But it doesn't do anything with those files yet.
This version also added a functionality where the app can read the *.db file from the external storage. So that the user can have a file in the external storage. The app then copies it to internal storage and displays the data.

+++ app/src/main/java/com/app/routineturboa/data/local/DatabaseHelper.kt
@@ -3,8 +3,10 @@ package com.app.routineturboa.data.local
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
+import android.os.Environment
 import android.provider.BaseColumns
 import java.io.File
+import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.IOException
 import java.io.InputStream
@@ -66,9 +68,19 @@ class DatabaseHelper(private val context: Context) : SQLiteOpenHelper(
         val dbPath: String = context.getDatabasePath(DATABASE_NAME).absolutePath
         val dbFile = File(dbPath)
         if (!dbFile.exists()) {
-            context.assets.open(DATABASE_NAME).use { inputStream: InputStream ->
-                FileOutputStream(dbPath).use { outputStream: OutputStream ->
-                    copyStream(inputStream, outputStream)
+            val externalDbPath = Environment.getExternalStorageDirectory().absolutePath + "/$DATABASE_NAME"
+            val externalDbFile = File(externalDbPath)
+            if (externalDbFile.exists()) {
+                FileInputStream(externalDbFile).use { inputStream: InputStream ->
+                    FileOutputStream(dbPath).use { outputStream: OutputStream ->
+                        copyStream(inputStream, outputStream)
+                    }
+                }
+            } else {
+                context.assets.open(DATABASE_NAME).use { inputStream: InputStream ->
+                    FileOutputStream(dbPath).use { outputStream: OutputStream ->
+                        copyStream(inputStream, outputStream)
+                    }
                 }
             }
         }
